Fully Implemented CommandAPI
